TASK-48123: Display pasted styled tables in notes (#193)
The problem was in the WikiHTMLSanitizer when the notes body contains a comment and just below a closing tag. After splitting the body, the second part will contain closing tags without their opening ones. So in the HTMLSanitizer.sanitize() method, they will be removed.
This process was done with the old wiki app to conserve the xwiki content in comments. We no longer need it in the new notes app.

This fix should replace the `markupSanitize` with the `HTMLSanitizer` to remove useless html comments.